+++ ntd-ecomerce-api/internal/usecase/mock_test.go
@@ -18,7 +18,7 @@ func (m *MockProductRepository) Add(_ context.Context, product domain.Product) (
 	return args.Get(0).(domain.Product), args.Error(1)
 }
 
-func (m *MockProductRepository) FindAll(_ context.Context, filter domain.ProductFilter, page domain.Page) (domain.ProductList, error) {
+func (m *MockProductRepository) FindAll(_ context.Context, filter domain.ProductFilter, page domain.PageRequest) (domain.ProductList, error) {
 	args := m.Called(filter, page)
 	return args.Get(0).(domain.ProductList), args.Error(1)
 }